requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7

# AI and Machine Learning
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0

# OLED Display dependencies
adafruit-circuitpython-ssd1306==2.12.14
adafruit-blinka==8.22.2
Pillow==10.1.0

# Additional utilities
requests==2.31.0
numpy==1.24.4

# Optional: For better performance
accelerate==0.24.1
